refactor(plugin): extract shared/wire.rs (slice_from_raw + write_err + write_id_bytes)
First content move under #1079. Lifts the three wire-format helpers
every host-side vtable callback uses out of host_services/mod.rs into
a new shared submodule. No call sites change — visibility widens from
file-private to pub(in crate::core::plugin::host_services) so future
sibling submodules can reach them via `super::shared::wire::*`.

Behavior identical; cargo test -p streamlib-engine --lib unchanged at
1407 passed / 2 pre-existing Deno failures / 126 ignored.
