+++ libs/streamlib-engine/src/core/plugin/host_services/shared/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) 2025 Jonathan Fontanez
+// SPDX-License-Identifier: BUSL-1.1
+
+//! Cross-submodule helpers used by every host-side vtable callback.
+
+pub(in crate::core::plugin::host_services) mod wire;
